requirements.txt
fastapi
uvicorn
openai
google-cloud-vision 
pdf2image 
poppler-utils